Use the component folder pattern and refactor FaCC
The <Home> component has other bits that won't be used anywhere else (like the
rules and hero image) so it makes sense to group these together.
The <App> component will likely follow suit once it gets filled out.

A reference to the component folder pattern is here:
https://medium.com/styled-components/component-folder-pattern-ee42df37ec68<Paste>

Also, although that article recommends against Function as Child Components:
http://americanexpress.io/faccs-are-an-antipattern/#component-injection---a-better-solution

It seems to me that the biggest argument against FaCC is that the naming is
confusing because you wouldn't expect the `children` prop to be a function
however I think this commit shows a pretty simple remedy to the naming problem
by simply naming it `render` when we destructure children.