Add player history and next/prev player events
This commit has a bit too much going on, but here is a rough list:
- Refactor PlayerCard
- Add timing debug info to the console for reducing events
- Add colorful history to PlayerCards
- Add logic for next and previous player events

The PlayerCard refactoring should hopefully allow me to export various
configurations of that components, like variations for:
- A small tile
- A big banner for the top
- A row entry
- etc.

The timing debugging is just to make sure the app is still responsive enough
with all the extra events for next/previous player actions.
I chose to go this way because even while it is still prone to human error,
that stats around "longest break" will still be possible (unless someone changes
the player by accident).

This also allows the remote control functionality straight away. If I want to
allow a user to look around without affecting the others, this can be achieved
by looking at the list of players on the screen (or in a side menu) or I'll have
to implement local state for that in addition to the shared state via firebase -
this sounds like more effort than its worth right now.