Use two different <PlayerCard> elements
The main <PlayerCard> is one that grows to fit the width while the
<SmallPlayerCard> has a fixed width.

A fixed width is important for the "other" players because you'll see them all
next to each other and it looks weird if they have different widths.
The automatic width is important for the current player because it should take
up more space to stand out more.