Prevent scoring unless opening with a cannon
This is to satisfy the rule of "getting off the mark":

> While your score is on zero, you can not begin to score any points until you
> successfully cannon off the pink - this will result in 2 points and you may
> continue to score using any other method after that.

Refactored the tests to only test what is required instead of everything!
There is still work to be done on this front I think, these tests are probably
still doing too much... and there are a lot of them!

There is also that bit "test everything" test at the end