Add 'undo' and refactor how scores are calculated
SCORING
-------

In order to implement 'undo' I decided to change how scores are calculated.

Instead of receiving a 'SCORE' or 'RESET' event and either adding to or
resetting the score in state, now it adds the history item to the history and
then reduces that down to a score instead.

While this means that events get reduced to history items AND for the history
gets reduced down to a score and this may seem to be adding exponential
complexity, it does mean the 'undo' feature just has to add an 'undo' to the
history and recalculate the score, rather than writing the logic to actually
reverse the previous operation ('SCORE' or 'RESET').

In my dev environment this nested `reduce` operation didn't seem to add any
noticeable delay.

The next step is to reduce some of the duplication in this implementation!!

REDO
----

Also, I decided not to implement 'redo' for now (maybe in the future) because:
- if you want to redo, you can just repeat the operation that you undid
- if you forget, the event log can be a reminder of which operation you need
- with the implementation before this, it would have been more difficult

Originally the idea was to make each item in the history array an object (instead
of an integer) that contains an "undone" boolean property, however this makes
the implementation a little tricky - consider the edge case below!

```
describe('redo', () => {
  it('removes the "UNDONE" mark from the last history item')
  it('has no effect if the last item is not "UNDONE"')
  // Consider events: 2,2,7,u,u,r,6,u,r
  it('takes previous undos into consideration')
})
```

(Note to self: check notes in my Samsung notebook)
Another option for 'redo' could be to create a 'redo' history array for each
player.