Refactor AddPlayerButton to be dialog-only FaCC
The AddPlayerButton component had two main jobs (and components):
1. A styled button
2. A dialog prompt that communicated with Firebase

When building the mobile view, I thought a floating action button would be
better and this means I'd need 2 variations of this AddPlayerButton component.

By changing this into a FaCC, this new AddPlayerDialog is really just the Dialog
with the Firebase logic, which is all invisible until you call a function.
This allows me to create whatever component I want around the app and re-use the
functionality of the Dialog and Firebase stuff without being coupled to a
particular button design, etc.