Make the UI reflect the available scoring options
While the reducer logic will not add to the score for non-cannons if the current
score is zero, the buttons still work and still send events.

While I don't specifically need to duplicate that logic here in the UI, having
the buttons disabled (and preventing the ripple animation) seems more correct
and it will save (incredibly small amounts of) firebase quota for mistakes.

... just being picky