Only listen for new documents since last update
This is an attempt as fixing issue #15 although time will tell if this issue
needs more work.

This change is a bit complicated but here is how I understand it:

Before this change
------------------

Given a collection with 100 documents:

1. Setup listener for ALL documents in collection
2. Listener is immediately called with ALL documents (100 reads)
3. Add a new document to the collection
4. The listener is called again with ALL documents in the collection (101 reads)

This results in 201 reads (2N+1) because every document that gets added triggers
a read for every other document!

After this change
-----------------

Given a collection with 100 documents:

1.  Setup listener for ALL documents in collection
2.  Listener is immediately called with ALL documents (100 reads)
3.  Store a reference to the "last document"
4.  Detach the listener
5.  Attach a new listener for ONLY documents after that "last document"
6.  Listener is immediately called with NO documents (0 reads)
7.  Add a new document to the collection
8.  Listener is called with ONLY new documents (1 read)
9.  Store a reference to the "last document"
10. Detach the listener
11. Attach a new listener for ONLY documents after that "last document"
12. Listener is immediately called with NO documents (0 reads)

This results in 101 reads (N) which should be a bit better on my quota!

An Alternative Approach
-----------------------

If this still doesn't help avoiding hitting the quota, another idea would be to
not store each event in Firebase but just process events locally and store the
resulting document to Firebase.

While this means the full event history would not saved in Firebase (which would
allow very detailed analysis of historic events) the history array of each
player would still be available and would provide most of the benefit.

In this modal, there is a risk of deleting the data and not being able to replay
the events as the current model can but this would decrease the amount of
firebase activity if needed.