Add snapshot tests for the 3 PlayerCard variants
The jest-styled-components package makes the error messages for failing tests
show the difference in CSS code in snapshots rather than simply an autogenerated
classname difference.

These tests have been added with the hope that I can refactor these different
variants into a single component that can be configured in a similar way to how
the current beta of material-ui does it (via a `variant` prop).