Collapse the different <PlayerCard>s into one
There was a lot of duplication around the different PlayerCard components and it
was fairly confusing to follow through all the different types.

The upside is that now the code is a simpler collection of components that can
probably be read and understood much easier and the <Card> component is now
called <ProgressPaper> which is more appropriate.

The downside is there are little conditionals everywhere. While I've never used
snapshot testing before this, I think this is where that type of testing is most
useful as writing unit tests around those may be a little tricky (at least with
the limited knowledge on the topic I currently have).

I think something like storybook would compliment this so I'd be able to
visually look at all the different variants easily without having to add players
through the UI, be connected to the internet for firebase data, etc. then if
any of those particular snapshot tests fail, I could look at the storybook and
probably see what has changed and judge if it was intentional or not.

This was part of the checklist of things in issue #7