Do not automatically shuffle upon end of game
As part of issue #27, this automatic shuffling has implications around
repeatable randomness.

The removal of this feature will mean the UI will need to allow shuffling to
happen and at the end of the game the user will have to press that button OR the
UI can automatically emit two events at the end of the game.