Don't put GAME_OVER and WIN in the history array
Instead, keep a 'wins' key on the player object. This will help with the
implementation of #14 and alleviate some trouble with issue #22.

This also makes a lot more sense from a user perspective because old games don't
really matter for the UI at that point in time and with this old implementation
you couldn't tell which score was from which game.

Fixes #18