Make the title go back to the homepage
This code was not great to start with and I think mixing in some inline styles
with the already out-of-the-ordinary classes object has made this worse!

But for now, this change fixes #30 and the quality of the code here should be
improved during #23 or next time something needs to be changed in here.