Move cumulativeScore calculation to reducer
The logic of calculating the cumulativeScore in the ChartPanel component
required logic from the reducer.
For this component to depend on and import from the reducer logic seemed wrong
and this logic didn't have any tests (I lazily tested this in the repl)

By moving it into the reducer, all game logic is once again in a single place,
it is now tested fairly thoroughly and the UI is simple again, it just takes the
state and renders out HTML, etc.

As part of this refactor I was also able to keep the calculateScore and
calculateCumulativeScore functions from being imported in reducers file because
it is only used by the newStateWithUpdatedPlayer function now and can
essentially be private which means they don't need to be exported now!