Add a share button with a QR code
This offers an easy way of starting a game on the laptop and then getting
everyone to open that same session on their phones.

I've moved the button components to fall under the AppBar because the App
component was getting pretty heavy and this doesn't really fit there although
this does mean the app is now given the currentPlayer in order for the
UndoButton to emit the correct event. This may need some refacting in the future
but it fixes #26 for now.

A simpler session id is also an option as detailed in issue #26:

> An alternative idea could be to generate a short lived, short code for the game.
>
> For example, upon a new game being created:
>
> create generated 20 char id as usual
> additionally create a 4 char short code that maps to the 20 char id
> use a simple URL like /join/12ab to redirect to the real URL with the 20 char id
> To clean these short codes up, some options would be:
>
> deleted upon access (single use)
> periodic firebase function to clean up
> stored with timestamp and upon creation all old ones are deleted